+++ internal/service/ec2/testdata/TransitGatewayMeteringPolicy/region_override/main_gen.tf
@@ -1,0 +1,19 @@
+# Copyright IBM Corp. 2014, 2026
+# SPDX-License-Identifier: MPL-2.0
+
+resource "aws_ec2_transit_gateway_metering_policy" "test" {
+  region = var.region
+
+  transit_gateway_id = aws_ec2_transit_gateway.test.id
+}
+
+resource "aws_ec2_transit_gateway" "test" {
+  region = var.region
+
+}
+
+variable "region" {
+  description = "Region to deploy resource in"
+  type        = string
+  nullable    = false
+}