+++ internal/service/events/target_list.go
@@ -94,7 +94,7 @@ func (l *listResourceTarget) List(ctx context.Context, request list.ListRequest,
 
 			result.DisplayName = targetID
 
-			l.SetResult(ctx, l.Meta(), request.IncludeResource, &result, rd)
+			l.SetResult(ctx, l.Meta(), request.IncludeResource, rd, &result)
 			if result.Diagnostics.HasError() {
 				yield(result)
 				return