+++ internal/service/s3/bucket_policy_list.go
@@ -11,7 +11,9 @@ import (
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 	awstypes "github.com/aws/aws-sdk-go-v2/service/s3/types"
+	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/list"
+	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-provider-aws/internal/errs/fwdiag"
 	"github.com/hashicorp/terraform-provider-aws/internal/framework"
@@ -24,61 +26,29 @@ import (
 // Function annotations are used for list resource registration to the Provider. DO NOT EDIT.
 // @SDKListResource("aws_s3_bucket_policy")
 func newBucketPolicyResourceAsListResource() inttypes.ListResourceForSDK {
-	l := listResourceBucketPolicy{}
-	l.SetResourceSchema(resourceBucketPolicy())
-	return &l
+	return newListResourceBaseBucketPropertySDK(
+		resourceBucketPolicy(),
+		newBucketPolicyListHandler,
+	)
 }
 
-var _ list.ListResource = &listResourceBucketPolicy{}
+var _ bucketPropertyListHandlerSDK = bucketPolicyListHandler{}
 
-type listResourceBucketPolicy struct {
-	framework.ListResourceWithSDKv2Resource
-}
-
-func (l *listResourceBucketPolicy) List(ctx context.Context, request list.ListRequest, stream *list.ListResultsStream) {
-	var query listBucketPolicyModel
-	if request.Config.Raw.IsKnown() && !request.Config.Raw.IsNull() {
-		if diags := request.Config.Get(ctx, &query); diags.HasError() {
-			stream.Results = list.ListResultsStreamDiagnostics(diags)
-			return
-		}
-	}
-
-	tflog.Info(ctx, "Listing S3 Bucket Policy")
-	stream.Results = func(yield func(list.ListResult) bool) {
-		tflog.Info(ctx, "Listing General Purpose Buckets")
-		gpConn := l.Meta().S3Client(ctx)
-		gpInput := s3.ListBucketsInput{
-			BucketRegion: aws.String(l.Meta().Region(ctx)),
-			MaxBuckets:   aws.Int32(int32(request.Limit)),
-		}
-		var count int64
-		for result := range l.list(ctx, request, gpConn, listBuckets(ctx, gpConn, &gpInput)) {
-			count++
-			if !yield(result) {
-				return
-			}
-		}
-
-		limit := request.Limit - count
-		if limit <= 0 {
-			tflog.Info(ctx, "Limit reached, skipping Directory Buckets")
-		}
-
-		tflog.Info(ctx, "Listing Directory Buckets")
-		dirConn := l.Meta().S3ExpressClient(ctx)
-		dirInput := s3.ListDirectoryBucketsInput{
-			MaxDirectoryBuckets: aws.Int32(int32(limit)),
-		}
-		for result := range l.list(ctx, request, dirConn, listDirectoryBuckets(ctx, dirConn, &dirInput)) {
-			if !yield(result) {
-				return
-			}
-		}
+func newBucketPolicyListHandler(lister listResourceSDK) bucketPropertyListHandlerSDK {
+	return bucketPolicyListHandler{
+		baseBucketPropertyListHandlerSDK: newBaseBucketPropertyListHandlerSDK(lister),
 	}
 }
 
-func (l *listResourceBucketPolicy) list(ctx context.Context, request list.ListRequest, conn *s3.Client, buckets iter.Seq2[awstypes.Bucket, error]) iter.Seq[list.ListResult] {
+type bucketPolicyListHandler struct {
+	baseBucketPropertyListHandlerSDK
+}
+
+func (l bucketPolicyListHandler) parseQuery(ctx context.Context, config tfsdk.Config) (diags diag.Diagnostics) {
+	return parseQuery[listBucketPolicyModel](ctx, config)
+}
+
+func (l bucketPolicyListHandler) list(ctx context.Context, request list.ListRequest, conn *s3.Client, buckets iter.Seq2[awstypes.Bucket, error]) iter.Seq[list.ListResult] {
 	return func(yield func(list.ListResult) bool) {
 		for bucket, err := range buckets {
 			if err != nil {
@@ -119,7 +89,7 @@ func (l *listResourceBucketPolicy) list(ctx context.Context, request list.ListRe
 
 			result.DisplayName = bucketName
 
-			l.SetResult(ctx, l.Meta(), request.IncludeResource, &result, rd)
+			l.SetResult(ctx, l.Meta(), request.IncludeResource, rd, &result)
 			if result.Diagnostics.HasError() {
 				yield(result)
 				return